requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.2
joblib==1.5.2
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
